NEX-1573 Split login screen into multiple components
